# api/generated.go
//go:build go1.22

// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package api

import (
	"context"
	"fmt"
	"net/http"

	"github.com/oapi-codegen/runtime"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Defines values for SettingSource.
const (
	SettingSourceDb      SettingSource = "db"
	SettingSourceDefault SettingSource = "default"
	SettingSourceEnv     SettingSource = "env"
)

// Defines values for SettingDetailSource.
const (
	SettingDetailSourceDb      SettingDetailSource = "db"
	SettingDetailSourceDefault SettingDetailSource = "default"
	SettingDetailSourceEnv     SettingDetailSource = "env"
)

// AuditEntry defines model for AuditEntry.
type AuditEntry struct {
	Action     *string `json:"action,omitempty"`
	Actor      *string `json:"actor,omitempty"`
	CreatedAt  *string `json:"created_at,omitempty"`
	Details    *string `json:"details,omitempty"`
	TargetId   *string `json:"target_id,omitempty"`
	TargetType *string `json:"target_type,omitempty"`
}

// ErrorEnvelope defines model for ErrorEnvelope.
type ErrorEnvelope struct {
	Error string `json:"error"`
	Ok    bool   `json:"ok"`
}

// Setting defines model for Setting.
type Setting struct {
	Default         *string        `json:"default,omitempty"`
	Key             *string        `json:"key,omitempty"`
	RuntimeEditable *bool          `json:"runtime_editable,omitempty"`
	Source          *SettingSource `json:"source,omitempty"`
	ValidValues     *[]string      `json:"valid_values,omitempty"`
	Value           *string        `json:"value,omitempty"`
}

// SettingSource defines model for Setting.Source.
type SettingSource string

// SettingDetail defines model for SettingDetail.
type SettingDetail struct {
	Default         *string              `json:"default,omitempty"`
	Description     *string              `json:"description,omitempty"`
	Key             *string              `json:"key,omitempty"`
	RuntimeEditable *bool                `json:"runtime_editable,omitempty"`
	Source          *SettingDetailSource `json:"source,omitempty"`
	ValidValues     *[]string            `json:"valid_values,omitempty"`
	Value           *string              `json:"value,omitempty"`
}

// SettingDetailSource defines model for SettingDetail.Source.
type SettingDetailSource string

// SuccessEnvelope defines model for SuccessEnvelope.
type SuccessEnvelope struct {
	Ok      bool                   `json:"ok"`
	Payload map[string]interface{} `json:"payload"`
}

// User defines model for User.
type User struct {
	Approved     *bool   `json:"approved,omitempty"`
	CreatedAt    *string `json:"created_at,omitempty"`
	Disabled     *bool   `json:"disabled,omitempty"`
	Id           *int64  `json:"id,omitempty"`
	IsAdmin      *bool   `json:"is_admin,omitempty"`
	LastLoginAt  *string `json:"last_login_at"`
	StorageQuota *int64  `json:"storage_quota,omitempty"`
	Username     *string `json:"username,omitempty"`
}

// Forbidden defines model for Forbidden.
type Forbidden = ErrorEnvelope

// Unauthorized defines model for Unauthorized.
type Unauthorized = ErrorEnvelope

// UnsupportedMediaType defines model for UnsupportedMediaType.
type UnsupportedMediaType = ErrorEnvelope

// AuditTailParams defines parameters for AuditTail.
type AuditTailParams struct {
	// N Number of entries to return
	N *int `form:"n,omitempty" json:"n,omitempty"`
}

// ConfigGetParams defines parameters for ConfigGet.
type ConfigGetParams struct {
	Key string `form:"key" json:"key"`
}

// ConfigResetJSONBody defines parameters for ConfigReset.
type ConfigResetJSONBody struct {
	Key string `json:"key"`
}

// ConfigSetJSONBody defines parameters for ConfigSet.
type ConfigSetJSONBody struct {
	Key   string `json:"key"`
	Value string `json:"value"`
}

// LoginJSONBody defines parameters for Login.
type LoginJSONBody struct {
	// Password Account password
	Password string `json:"password"`

	// Username Admin username
	Username string `json:"username"`
}

// LogoutJSONBody defines parameters for Logout.
type LogoutJSONBody = map[string]interface{}

// UserAddJSONBody defines parameters for UserAdd.
type UserAddJSONBody struct {
	IsAdmin  *bool  `json:"is_admin,omitempty"`
	Password string `json:"password"`
	Username string `json:"username"`
}

// UserDeleteJSONBody defines parameters for UserDelete.
type UserDeleteJSONBody struct {
	Username string `json:"username"`
}

// UserDisableJSONBody defines parameters for UserDisable.
type UserDisableJSONBody struct {
	Username string `json:"username"`
}

// UserPasswdJSONBody defines parameters for UserPasswd.
type UserPasswdJSONBody struct {
	Password string `json:"password"`
	Username string `json:"username"`
}

// UserApproveJSONBody defines parameters for UserApprove.
type UserApproveJSONBody struct {
	Username string `json:"username"`
}

// UserPromoteJSONBody defines parameters for UserPromote.
type UserPromoteJSONBody struct {
	Username string `json:"username"`
}

// UserRejectJSONBody defines parameters for UserReject.
type UserRejectJSONBody struct {
	Username string `json:"username"`
}

// ConfigResetJSONRequestBody defines body for ConfigReset for application/json ContentType.
type ConfigResetJSONRequestBody ConfigResetJSONBody

// ConfigSetJSONRequestBody defines body for ConfigSet for application/json ContentType.
type ConfigSetJSONRequestBody ConfigSetJSONBody

// LoginJSONRequestBody defines body for Login for application/json ContentType.
type LoginJSONRequestBody LoginJSONBody

// LogoutJSONRequestBody defines body for Logout for application/json ContentType.
type LogoutJSONRequestBody = LogoutJSONBody

// UserAddJSONRequestBody defines body for UserAdd for application/json ContentType.
type UserAddJSONRequestBody UserAddJSONBody

// UserDeleteJSONRequestBody defines body for UserDelete for application/json ContentType.
type UserDeleteJSONRequestBody UserDeleteJSONBody

// UserDisableJSONRequestBody defines body for UserDisable for application/json ContentType.
type UserDisableJSONRequestBody UserDisableJSONBody

// UserPasswdJSONRequestBody defines body for UserPasswd for application/json ContentType.
type UserPasswdJSONRequestBody UserPasswdJSONBody

// UserApproveJSONRequestBody defines body for UserApprove for application/json ContentType.
type UserApproveJSONRequestBody UserApproveJSONBody

// UserPromoteJSONRequestBody defines body for UserPromote for application/json ContentType.
type UserPromoteJSONRequestBody UserPromoteJSONBody

// UserRejectJSONRequestBody defines body for UserReject for application/json ContentType.
type UserRejectJSONRequestBody UserRejectJSONBody

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Fetch recent audit log entries
	// (GET /json/audit/tail)
	AuditTail(w http.ResponseWriter, r *http.Request, params AuditTailParams)
	// Get a single configuration setting
	// (GET /json/config/get)
	ConfigGet(w http.ResponseWriter, r *http.Request, params ConfigGetParams)
	// List all configuration settings
	// (GET /json/config/list)
	ConfigList(w http.ResponseWriter, r *http.Request)
	// Reset a setting to its default value
	// (POST /json/config/reset)
	ConfigReset(w http.ResponseWriter, r *http.Request)
	// Update a runtime-editable setting
	// (POST /json/config/set)
	ConfigSet(w http.ResponseWriter, r *http.Request)
	// Authenticate and obtain a session token
	// (POST /json/login)
	Login(w http.ResponseWriter, r *http.Request)
	// Destroy the current session
	// (POST /json/logout)
	Logout(w http.ResponseWriter, r *http.Request)
	// Create a new user
	// (POST /json/user/add)
	UserAdd(w http.ResponseWriter, r *http.Request)
	// Approve a pending user
	// (POST /json/user/approve)
	UserApprove(w http.ResponseWriter, r *http.Request)
	// Delete a user
	// (POST /json/user/delete)
	UserDelete(w http.ResponseWriter, r *http.Request)
	// Toggle disabled status
	// (POST /json/user/disable)
	UserDisable(w http.ResponseWriter, r *http.Request)
	// List all users
	// (GET /json/user/list)
	UserList(w http.ResponseWriter, r *http.Request)
	// Change a user's password
	// (POST /json/user/passwd)
	UserPasswd(w http.ResponseWriter, r *http.Request)
	// Toggle admin status
	// (POST /json/user/promote)
	UserPromote(w http.ResponseWriter, r *http.Request)
	// Reject a pending user
	// (POST /json/user/reject)
	UserReject(w http.ResponseWriter, r *http.Request)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// AuditTail operation middleware
func (siw *ServerInterfaceWrapper) AuditTail(w http.ResponseWriter, r *http.Request) {

	var err error

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	// Parameter object where we will unmarshal all parameters from the context
	var params AuditTailParams

	// ------------- Optional query parameter "n" -------------

	err = runtime.BindQueryParameter("form", true, false, "n", r.URL.Query(), &params.N)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "n", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.AuditTail(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ConfigGet operation middleware
func (siw *ServerInterfaceWrapper) ConfigGet(w http.ResponseWriter, r *http.Request) {

	var err error

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	// Parameter object where we will unmarshal all parameters from the context
	var params ConfigGetParams

	// ------------- Required query parameter "key" -------------

	if paramValue := r.URL.Query().Get("key"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "key"})
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "key", r.URL.Query(), &params.Key)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "key", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ConfigGet(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ConfigList operation middleware
func (siw *ServerInterfaceWrapper) ConfigList(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ConfigList(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ConfigReset operation middleware
func (siw *ServerInterfaceWrapper) ConfigReset(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ConfigReset(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ConfigSet operation middleware
func (siw *ServerInterfaceWrapper) ConfigSet(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ConfigSet(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// Login operation middleware
func (siw *ServerInterfaceWrapper) Login(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Login(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// Logout operation middleware
func (siw *ServerInterfaceWrapper) Logout(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Logout(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UserAdd operation middleware
func (siw *ServerInterfaceWrapper) UserAdd(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UserAdd(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UserApprove operation middleware
func (siw *ServerInterfaceWrapper) UserApprove(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UserApprove(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UserDelete operation middleware
func (siw *ServerInterfaceWrapper) UserDelete(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UserDelete(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UserDisable operation middleware
func (siw *ServerInterfaceWrapper) UserDisable(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UserDisable(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UserList operation middleware
func (siw *ServerInterfaceWrapper) UserList(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UserList(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UserPasswd operation middleware
func (siw *ServerInterfaceWrapper) UserPasswd(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UserPasswd(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UserPromote operation middleware
func (siw *ServerInterfaceWrapper) UserPromote(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UserPromote(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// UserReject operation middleware
func (siw *ServerInterfaceWrapper) UserReject(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.UserReject(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of http.ServeMux.
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	ServeHTTP(w http.ResponseWriter, r *http.Request)
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc("GET "+options.BaseURL+"/json/audit/tail", wrapper.AuditTail)
	m.HandleFunc("GET "+options.BaseURL+"/json/config/get", wrapper.ConfigGet)
	m.HandleFunc("GET "+options.BaseURL+"/json/config/list", wrapper.ConfigList)
	m.HandleFunc("POST "+options.BaseURL+"/json/config/reset", wrapper.ConfigReset)
	m.HandleFunc("POST "+options.BaseURL+"/json/config/set", wrapper.ConfigSet)
	m.HandleFunc("POST "+options.BaseURL+"/json/login", wrapper.Login)
	m.HandleFunc("POST "+options.BaseURL+"/json/logout", wrapper.Logout)
	m.HandleFunc("POST "+options.BaseURL+"/json/user/add", wrapper.UserAdd)
	m.HandleFunc("POST "+options.BaseURL+"/json/user/approve", wrapper.UserApprove)
	m.HandleFunc("POST "+options.BaseURL+"/json/user/delete", wrapper.UserDelete)
	m.HandleFunc("POST "+options.BaseURL+"/json/user/disable", wrapper.UserDisable)
	m.HandleFunc("GET "+options.BaseURL+"/json/user/list", wrapper.UserList)
	m.HandleFunc("POST "+options.BaseURL+"/json/user/passwd", wrapper.UserPasswd)
	m.HandleFunc("POST "+options.BaseURL+"/json/user/promote", wrapper.UserPromote)
	m.HandleFunc("POST "+options.BaseURL+"/json/user/reject", wrapper.UserReject)

	return m
}